Fixed Dart and Windows problems
