+++ lidar/analysis_options.yaml
@@ -1,4 +1,4 @@
-linter:
-  rules:
-    # Temporarily disabled until we are ready to document
-    public_member_api_docs: false
+# linter:
+#   rules:
+#     # Temporarily disabled until we are ready to document
+#     public_member_api_docs: false
