+++ lidar/lib/lidar.dart
@@ -1,3 +1,2 @@
 export "src/collection.dart";
 export "src/lidar.dart";
-export "src/lidar_stub.dart";
