+++ lidar/lib/src/collection.dart
@@ -1,9 +1,9 @@
 import "package:burt_network/burt_network.dart";
 
-import "lidar_stub.dart";
+import "lidar.dart";
 
 class LidarCollection extends Service {
-  final lidar = LidarStub();
+  final lidar = Lidar();
 
   //final server = RoverSocket(port: 8002, device: Device.VIDEO);
 
@@ -15,7 +15,7 @@ class LidarCollection extends Service {
   Future<bool> init() async {
      //status &= await server.init();
     status &= await lidar.init();
-    print("Inited with status $status");
+    print("Initialized with status $status");
     return status;
   }
